+++ UAV-Map-Converter/src/uav/map/converter/map/TypeRegion.java
@@ -1,0 +1,11 @@
+package uav.map.converter.map;
+
+/**
+ *
+ * @author Jesimar Arantes
+ */
+public enum TypeRegion {
+    
+    NFZ_REGION, BONUS_REGION, PENALTY_REGION;
+    
+}